Remove duplication in test files
