+++ test/trelloDOM.js
@@ -1,41 +1,12 @@
-var vm = require('vm');
 var fs = require('fs');
-var chrome = require('sinon-chrome');
+var vm = require('vm');
 var assert = require('assert');
-
-// Stop the DEV_MODE const from causing an error
-chrome.runtime.getManifest.returns([]);
-
-// Fake up a browser to stop calls that access the DOM from causing errors
-var MockBrowser = require('mock-browser').mocks.MockBrowser;
-var mock = new MockBrowser();
-var document = mock.getDocument();
-var window = mock.getWindow();
-var setInterval = function () {}
-
-// This is the context our tests will run in
-var sandbox = {
-  chrome,
-  document,
-  setInterval,
-  returnValue: null
-};
-var context = vm.createContext(sandbox);
-
-// Load the next-step-for-trello code
-var code = fs.readFileSync('next-step-for-trello.user.js')
-const script = new vm.Script(code);
-script.runInContext(context)
-
-// Mock the getUserName function
-var mockGetUserName = 'getUserName = function () { return "test"; }';
-const mockGetUserNameScript = new vm.Script(mockGetUserName);
-mockGetUserNameScript.runInContext(context)
+var test = require('./testEnvironment.js')
 
 // Load the mocked broswer document with the card HTML we expect to see
-var myDiv = document.createElement('div');
+var myDiv = test.document.createElement('div');
 myDiv.innerHTML = fs.readFileSync('test/expectedList.html');
-document.body.appendChild(myDiv);
+test.document.body.appendChild(myDiv);
 
 // Set up and run our test cases
 describe('getCardElementByShortUrl', function() {
@@ -43,16 +14,16 @@ describe('getCardElementByShortUrl', function() {
     var testScript = new vm.Script(
       'returnValue = getCardElementByShortUrl("https://trello.com/c/invalid")'
     );
-    testScript.runInContext(context);
-    assert.equal(sandbox.returnValue, undefined);
+    testScript.runInContext(test.context);
+    assert.equal(test.sandbox.returnValue, undefined);
   });
 
   it('returns card element if URL does exist', function() {
     var testScript = new vm.Script(
       'returnValue = getCardElementByShortUrl("https://trello.com/c/1234")'
     );
-    testScript.runInContext(context);
-    assert.equal(sandbox.returnValue instanceof window.HTMLSpanElement, true);
+    testScript.runInContext(test.context);
+    assert.equal(test.sandbox.returnValue instanceof test.window.HTMLSpanElement, true);
   });
 });
 
@@ -62,8 +33,8 @@ describe('getCardUrlFromTitleElement', function() {
       'var titleElement = document.getElementsByClassName("list-card-title")[0];'
       + 'returnValue = getCardUrlFromTitleElement(titleElement)'
     );
-    testScript.runInContext(context);
-    assert.equal(sandbox.returnValue, '/c/1234/1-test-1');
+    testScript.runInContext(test.context);
+    assert.equal(test.sandbox.returnValue, '/c/1234/1-test-1');
   });
 });
 
@@ -72,8 +43,8 @@ describe('elementIsTrelloCard', function () {
     var testScript = new vm.Script(
       'returnValue = elementIsTrelloCard(document.body)'
     );
-    testScript.runInContext(context);
-    assert.equal(sandbox.returnValue, false);
+    testScript.runInContext(test.context);
+    assert.equal(test.sandbox.returnValue, false);
   });
 
   it('returns true for inactive card', function () {
@@ -81,8 +52,8 @@ describe('elementIsTrelloCard', function () {
       'var card = document.getElementsByClassName("list-card")[0];'
       + 'returnValue = elementIsTrelloCard(card)'
     );
-    testScript.runInContext(context);
-    assert.equal(sandbox.returnValue, true);
+    testScript.runInContext(test.context);
+    assert.equal(test.sandbox.returnValue, true);
   });
 
   it('returns true for active card', function () {
@@ -90,8 +61,8 @@ describe('elementIsTrelloCard', function () {
       'var card = document.getElementsByClassName("list-card")[1];'
       + 'returnValue = elementIsTrelloCard(card)'
     );
-    testScript.runInContext(context);
-    assert.equal(sandbox.returnValue, true);
+    testScript.runInContext(test.context);
+    assert.equal(test.sandbox.returnValue, true);
   });
 
   it('returns true for card with cover', function () {
@@ -99,7 +70,7 @@ describe('elementIsTrelloCard', function () {
       'var card = document.getElementsByClassName("list-card")[2];'
       + 'returnValue = elementIsTrelloCard(card)'
     );
-    testScript.runInContext(context);
-    assert.equal(sandbox.returnValue, true);
+    testScript.runInContext(test.context);
+    assert.equal(test.sandbox.returnValue, true);
   });
 });
